% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scTrends-package.R
\docType{package}
\name{scTrends-package}
\alias{scTrends-package}
\title{scTrends: Trend detection for single-cell pseudotime analyses}
\description{
The \strong{scTrends} package provides a workflow to:
\enumerate{
\item Bin single-cell expression profiles along pseudotime using \code{\link[=scBin]{scBin()}}.
\item Automatically tune trend-detection parameters via \code{\link[=scPS]{scPS()}}.
\item Classify genes into temporal trend categories using \code{\link[=scTrends]{scTrends()}}.
\item Visualize gene sets or individual genes along pseudotime with
\code{\link[=plot_trends]{plot_trends()}} and \code{\link[=plot_genes]{plot_genes()}}.
}

The typical workflow is:
\preformatted{
  sb  <- scBin(seurat_obj, pseudotime_col = "pseudotime")
  ps  <- scPS(sb)
  tr  <- do.call(scTrends, c(list(scbin_result = sb), ps$optimal_params))

  # Plot all significant Up genes
  p1 <- plot_trend(tr, bin_means = sb$bin_means, x = sb$x,
                   trend_type = "Up", sig_metric = "q_value")

  # Plot a few marker genes
  p2 <- plot_genes(sb$bin_means, sb$x,
                   genes = c("WT1", "NPHS1", "NPHS2"))
}
}
\author{
\strong{Maintainer}: Jianbo Qing \email{Dr.jianboqing@email.com}

}
\keyword{internal}
